pyservice/requirements.txt
# =============================================================================
# PyService Mini-ITSM Platform - Python Dependencies
# =============================================================================

# Core Framework
Django>=4.2,<5.0
djangorestframework>=3.14

# Database
mysqlclient>=2.2

# =============================================================================
# Celery & Async Processing
# =============================================================================
celery>=5.3
redis>=5.0
django-celery-beat>=2.5
django-celery-results>=2.5

# =============================================================================
# Security & Authentication
# =============================================================================
djangorestframework-simplejwt>=5.3
django-ratelimit>=4.1
django-cors-headers>=4.3
pyotp>=2.9
qrcode>=7.4

# =============================================================================
# Production & Performance
# =============================================================================
gunicorn>=21.2
whitenoise>=6.6
python-decouple>=3.8

# =============================================================================
# Logging & Monitoring
# =============================================================================
python-json-logger>=2.0
django-prometheus>=2.3

# =============================================================================
# Real-time & WebSocket
# =============================================================================
channels>=4.0
channels-redis>=4.1
daphne>=4.0

# =============================================================================
# PDF & Reporting
# =============================================================================
reportlab>=4.0
Pillow>=10.0

# =============================================================================
# GraphQL
# =============================================================================
graphene-django>=3.1

# =============================================================================
# Search
# =============================================================================
elasticsearch>=8.11
elasticsearch-dsl>=8.11
django-elasticsearch-dsl>=8.0

# =============================================================================
# API Documentation
# =============================================================================
drf-spectacular>=0.27

# =============================================================================
# Testing
# =============================================================================
pytest>=8.0
pytest-django>=4.7
pytest-cov>=4.1
pytest-asyncio>=0.23
factory-boy>=3.3

# =============================================================================
# Code Quality
# =============================================================================
black>=23.12
flake8>=6.1
isort>=5.13
bandit>=1.7

# =============================================================================
# Utilities
# =============================================================================
python-dateutil>=2.8
